fixed the -negative input message'

+++ src/main/java/rocks/zipcode/atm/CashMachine.java
@@ -35,7 +35,7 @@ public class CashMachine {
         );
     }
 
-    public void deposit(int amount) {
+    public void deposit(Float amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.deposit(accountData, amount),
@@ -44,7 +44,7 @@ public class CashMachine {
         }
     }
 
-    public void withdraw(int amount) {
+    public void withdraw(Float amount) {
         if (accountData != null) {
              tryCall(
                     () -> bank.withdraw(accountData, amount),
